fix(play-history): cast ROW_NUMBER to int so grouped query JSON-serializes
The grouped activity-history query left ROW_NUMBER() OVER (...) uncast, so
Prisma returned `rn` as a JS BigInt; spreading it into the response made
NextResponse.json throw "Do not know how to serialize a BigInt" and 500 on
every default (grouped) /api/play-history load. Cast it ::int like the
sibling window aggregates.

+++ src/app/api/play-history/route.ts
@@ -273,7 +273,7 @@ async function groupedQuery(
         ROW_NUMBER() OVER (
           PARTITION BY COALESCE(h."referenceId", h.id)
           ORDER BY h."startedAt" DESC
-        ) AS rn,
+        )::int AS rn,
         COUNT(*) OVER (
           PARTITION BY COALESCE(h."referenceId", h.id)
         )::int AS segment_count,
